Publish Progress Steps Project live
Create HTML markup for the Progress bars,
Progress Steps and 'Prev' and 'Next' buttons.

Style the Progess bars, Progress steps and buttons
with animations to render the coloring effects
smooth.

Implement logic to color progressively the bars and
circles to blue once we click the 'Next' button, and
regressively once we click the 'Prev' button.